feat(dj_car_services): Create initial Django project structure
This commit introduces the base project structure for the 'dj_car_services' application developed in Django. The core Django settings file, URLs mapping file, ASGI and WSGI configuration are added. It also includes some boilerplate files like .gitignore, README, manage.py which are necessary for the application setup. README file provides a comprehensive introduction to the project and steps to run it at the local machine. The .gitignore file includes typical ignore rules for a Django project. The addition of these files lays the groundwork for further development.