refactor(vehicles-models): Use ColorEnum for Vehicle's color field
Changed the 'color' field of the 'Vehicle' model from a CharField to an IntegerField with the 'choices' attribute limited to values defined in 'ColorEnum', promoting consistency in representing 'Vehicle' colors throughout the application and enabling easier maintainability through future 'ColorEnum' updates. Accompanying this change, a new migration file is created for application to the database.