refactor(vehicles-admin): Use VehicleForm in VehicleAdmin
The 'VehicleForm' was imported and integrated into the 'VehicleAdmin' class in the vehicles/admin.py file. The adjustment enables 'VehicleAdmin' to utilize fields defined in 'VehicleForm', enhancing its adaptability and getting rid of the need for manual year selection function. 'VehicleForm' now efficiently generates a list of year choices, from the present year to 1970, in descending order.